Create README.md
This work aims to create a tool that automatically identifies ethnicity based on text. We train a deep learning model on a large dataset of text samples from various ethnic groups to achieve this. So the main goal is to develop a tool that can accurately define text samples into specified ethnic categories by training a model on a broad dataset that includes different ethnic groups and associated linguistic features